futurize --stage2 -w -n

+++ lib/python/Tools/StbHardware.py
@@ -1,4 +1,5 @@
 from __future__ import print_function
+from builtins import str
 from time import time, localtime, gmtime
 from os import path
 from fcntl import ioctl
@@ -16,7 +17,7 @@ def getFPVersion():
 		elif getBoxType() in ('dm7080','dm820','dm520','dm525','dm900','dm920'):
 			ret = open("/proc/stb/fp/version", "r").read()
 		else:	
-			ret = long(open("/proc/stb/fp/version", "r").read())
+			ret = int(open("/proc/stb/fp/version", "r").read())
 	except IOError:
 		try:
 			fp = open("/dev/dbox/fp0")
@@ -65,7 +66,7 @@ def getFPWakeuptime():
 	ret = 0
 	try:
 		f = open("/proc/stb/fp/wakeup_time", "r")
-		ret = long(f.read())
+		ret = int(f.read())
 		f.close()
 	except IOError:
 		try:
